refactor: Service, UseCase 계층 Convert, Extract 메서드 별도 클래스로 분리
