+++ src/main/java/online/bottler/mapletter/application/MapLetterProximityService.java
@@ -2,14 +2,12 @@ package online.bottler.mapletter.application;
 
 import java.math.BigDecimal;
 import java.util.List;
-import java.util.Map;
 import java.util.Set;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import online.bottler.mapletter.application.dto.MapLetterAndDistance;
 import online.bottler.mapletter.application.port.in.MapLetterProximityUseCase;
 import online.bottler.mapletter.application.port.out.MapLetterPersistencePort;
-import online.bottler.mapletter.application.response.FindNearbyLettersResponse;
 import online.bottler.mapletter.domain.MapLetter;
 import online.bottler.mapletter.domain.policy.MapLetterPolicy;
 import org.springframework.stereotype.Service;
@@ -30,27 +28,9 @@ public class MapLetterProximityService implements MapLetterProximityUseCase {
     }
 
     @Override
-    public Set<Long> extractCreateUserIds(List<MapLetterAndDistance> letters) {
-        return letters.stream()
-                .map(MapLetterAndDistance::getCreateUserId)
-                .collect(Collectors.toSet());
-    }
-
-    @Override
     @Transactional(readOnly = true)
     public List<MapLetterAndDistance> findGuestNearByMapLetters(BigDecimal latitude, BigDecimal longitude) {
         return mapLetterPersistencePort.guestFindLettersByUserLocation(latitude, longitude);
-    }
-
-    @Override
-    public List<FindNearbyLettersResponse> convertToFindNearbyResponses(List<MapLetterAndDistance> letters,
-                                                                        Map<Long, String> nicknamesByIds) {
-        return letters.stream()
-                .map(letter -> {
-                    String nickname = nicknamesByIds.get(letter.getCreateUserId());
-                    return FindNearbyLettersResponse.from(letter, nickname);
-                })
-                .toList();
     }
 
     @Override
